test(architecture): move infrastructure production smoke tests
