+++ test/unit/infrastructure/env.test.ts
@@ -10,7 +10,7 @@ process.env.SUPABASE_SERVICE_ROLE_KEY ??= "test-service-role-key";
 process.env.DATABASE_URL ??= "postgresql://postgres:postgres@127.0.0.1:5432/postgres";
 process.env.SUPABASE_DB_URL ??= process.env.DATABASE_URL;
 
-const { ENV, resolvePublicSiteUrl } = await import("../server/lib/env.ts");
+const { ENV, resolvePublicSiteUrl } = await import("../../../server/lib/env.ts");
 
 function readGmailApiEnvFromChild(overrides: Record<string, string>) {
   const env = {
