+++ test/unit/infrastructure/package-scripts-contract.test.ts
