+++ test/unit/infrastructure/progress-production-invariants.test.ts
