+++ test/unit/infrastructure/smoke-local-contract.test.ts
